+++ genmeta-nat/src/main.rs
@@ -1,6 +1,6 @@
 use clap::Parser;
 
-#[tokio::main(flavor = "current_thread")]
+#[tokio::main]
 async fn main() {
     tracing_subscriber::fmt()
         .with_env_filter(