+++ genmeta-ssh3/src/main.rs
@@ -1,16 +1,7 @@
 use clap::Parser;
 
-#[tokio::main(flavor = "current_thread")]
+#[tokio::main]
 async fn main() {
-    tracing_subscriber::fmt()
-        .with_env_filter(
-            tracing_subscriber::EnvFilter::builder()
-                .with_default_directive(tracing_subscriber::filter::LevelFilter::OFF.into())
-                .from_env_lossy(),
-        )
-        .with_writer(std::io::stderr)
-        .init();
-
     if let Err(error) = genmeta_ssh3::run(genmeta_ssh3::Options::parse()).await {
         eprintln!("{error}");
         tracing::error!("Exit with error: {}", error);