+++ ssh-config/src/lib.rs
@@ -1,0 +1,9 @@
+pub mod ast;
+pub mod error;
+pub mod pattern;
+
+#[cfg(feature = "openssh")]
+pub mod openssh;
+
+#[cfg(feature = "genmeta")]
+pub mod genmeta;