+++ CHANGELOG.md
@@ -5,12 +5,27 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [0.4.0] - 2025-9-?
+
+### Changed
+- 新增genmeta-doctor，genmeta-nat 移入其中作为net子命令
+- 新增ssh-config crate用于处理openssh和genmeta配置文件（ssh config语法）
+- genmeta-ssh3 支持--id参数携带身份，解析profile身份配置文件，发起带有client name参数和tls证书的quic连接
+
+### Components
+- genmeta v0.4.0 ?
+- genmeta-ssh3 v0.4.0 ?
+- genmeta-curl v0.2.0 ?
+- genmeta-doctor v0.1.0
+- ssh-config v0.1.0
+
 ## [0.3.2] - 2025-7-30
 
-## Changed
+### Changed
 - ssh3支持了windows平台
 
-## Components
+### Components
 - genmeta v0.3.2
 - genmeta-ssh3 v0.3.1
 
@@ -62,7 +77,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.2.7] - 2025-06-11
 
 ### Added
-- ssh3, request, nslookup 支持使用~省略.genemta.net
+- ssh3, request, nslookup 支持使用~省略.genmeta.net
 
 ### Changed
 - 更新依赖，提升打洞能力